requirements.txt
adjustText==0.7.3
anndata==0.8.0
backcall==0.2.0
click==8.1.3
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
DensityPlot==0.1.8
dunamai==1.14.1
entrypoints==0.4
fonttools==4.38.0
get_version==3.5.4
h5py==3.7.0
importlib-metadata==5.1.0
ipykernel==6.16.2
ipython==7.34.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.7
jupyter_core==4.11.2
kiwisolver==1.4.4
legacy-api-wrap==1.2
llvmlite==0.31.0
loompy==3.0.7
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
matplotlib-venn==0.11.7
natsort==8.2.0
nest-asyncio==1.5.6
networkx==2.6.3
nheatmap==0.1.4
numba==0.48.0
numexpr==2.8.4
numpy==1.21.6
numpy-groupies==0.9.20
packaging==21.3
pandas==1.1.5
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==22.3.1
plotly==5.11.0
prompt-toolkit==3.0.33
psutil==5.9.4
ptyprocess==0.7.0
Pygments==2.13.0
pynndescent==0.5.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
pyzmq==24.0.1
sam-algorithm==0.7.1
scanpy==1.6.0
scikit-learn==1.0.2
scipy==1.7.3
scvelo==0.1.25
seaborn==0.12.1
session-info==1.0.0
setuptools==65.5.1
setuptools-scm==7.0.5
sinfo==0.3.4
six==1.16.0
statsmodels==0.13.5
stdlib-list==0.8.0
tables==3.7.0
tbb==2021.7.1
tenacity==8.1.0
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
umap-learn==0.3.10
wcwidth==0.2.5
wheel==0.38.4
zipp==3.11.0
